# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BarangayConnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  email: string;
  /** @minLength 6 */
  password: string;
  address: string;
  contactNumber: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  resident: 'resident',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  disabled: 'disabled',
} as const;

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserList {
  data: User[];
  total: number;
  page: number;
  limit: number;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  resident: 'resident',
} as const;

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  address?: string;
  contactNumber?: string;
  role?: UserUpdateRole;
}

export interface ProfileUpdate {
  firstName?: string;
  lastName?: string;
  address?: string;
  contactNumber?: string;
  /** @minLength 6 */
  password?: string;
}

export type RequestType = typeof RequestType[keyof typeof RequestType];


export const RequestType = {
  barangay_clearance: 'barangay_clearance',
  certificate_of_residency: 'certificate_of_residency',
  business_permit: 'business_permit',
  complaint: 'complaint',
  community_concern: 'community_concern',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];


export const RequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Request {
  id: number;
  userId: number;
  user?: User;
  type: RequestType;
  status: RequestStatus;
  subject: string;
  description: string;
  /** @nullable */
  remarks?: string | null;
  /** @nullable */
  processedBy?: number | null;
  /** @nullable */
  processedAt?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface RequestList {
  data: Request[];
  total: number;
  page: number;
  limit: number;
}

export type RequestInputType = typeof RequestInputType[keyof typeof RequestInputType];


export const RequestInputType = {
  barangay_clearance: 'barangay_clearance',
  certificate_of_residency: 'certificate_of_residency',
  business_permit: 'business_permit',
  complaint: 'complaint',
  community_concern: 'community_concern',
} as const;

export interface RequestInput {
  type: RequestInputType;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  description: string;
}

export interface RequestUpdate {
  subject?: string;
  description?: string;
}

export interface RequestAction {
  remarks?: string;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  category?: string | null;
  authorId: number;
  author?: User;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface AnnouncementList {
  data: Announcement[];
  total: number;
  page: number;
  limit: number;
}

export interface AnnouncementInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  category?: string;
}

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  category?: string;
}

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  /** @nullable */
  type?: string | null;
  isRead: boolean;
  /** @nullable */
  requestId?: number | null;
  createdAt: string;
}

export interface NotificationList {
  data: Notification[];
  unreadCount: number;
}

export interface DashboardStats {
  totalResidents: number;
  totalRequests: number;
  pendingRequests: number;
  approvedRequests: number;
  rejectedRequests: number;
  totalAnnouncements: number;
  recentActivity: number;
}

export interface BreakdownItem {
  label: string;
  count: number;
}

export interface RequestBreakdown {
  byType: BreakdownItem[];
  byStatus: BreakdownItem[];
}

export type ListUsersParams = {
search?: string;
role?: ListUsersRole;
status?: ListUsersStatus;
page?: number;
limit?: number;
};

export type ListUsersRole = typeof ListUsersRole[keyof typeof ListUsersRole];


export const ListUsersRole = {
  admin: 'admin',
  resident: 'resident',
} as const;

export type ListUsersStatus = typeof ListUsersStatus[keyof typeof ListUsersStatus];


export const ListUsersStatus = {
  active: 'active',
  disabled: 'disabled',
} as const;

export type ListRequestsParams = {
status?: ListRequestsStatus;
type?: ListRequestsType;
userId?: number;
search?: string;
page?: number;
limit?: number;
};

export type ListRequestsStatus = typeof ListRequestsStatus[keyof typeof ListRequestsStatus];


export const ListRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type ListRequestsType = typeof ListRequestsType[keyof typeof ListRequestsType];


export const ListRequestsType = {
  barangay_clearance: 'barangay_clearance',
  certificate_of_residency: 'certificate_of_residency',
  business_permit: 'business_permit',
  complaint: 'complaint',
  community_concern: 'community_concern',
} as const;

export type ListMyRequestsParams = {
status?: ListMyRequestsStatus;
page?: number;
limit?: number;
};

export type ListMyRequestsStatus = typeof ListMyRequestsStatus[keyof typeof ListMyRequestsStatus];


export const ListMyRequestsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type ListAnnouncementsParams = {
page?: number;
limit?: number;
};

export type ListNotificationsParams = {
unreadOnly?: boolean;
};

export type GetRecentRequestsParams = {
limit?: number;
};