# lib/api-zod/src/generated/types/announcementUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BarangayConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  category?: string;
}