# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BarangayConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  createdAt: Date;
  /** @nullable */
  updatedAt?: string | null;
}